reuse getTokenMetadata for registry and tokenlists
